Refactor DataProvider by extracting out the search implementation, logger, and protocol into separate sources, remove NetworkKit as a dependancy from PokedexCommon
