+++ Pokedex/Project.swift
@@ -19,7 +19,7 @@ let project = Project.app(name: "Pokedex",
                                                              resources: []),
                                               LocalFramework(name: "PokedexCommon",
                                                             path: "Pokedex",
-                                                            frameworkDependancies: [.target(name: "NetworkKit")],
+                                                            frameworkDependancies: [],
                                                             resources: ["Targets/PokedexCommon/Sources/**/*.xib"]),
                                               LocalFramework(name: "NetworkKit",
                                                              path: "Network",
