+++ Pokedex/Targets/PokedexCommon/Sources/Core/DataProviding.swift
@@ -1,0 +1,15 @@
+//
+//  DataProviding.swift
+//  PokedexCommon
+//
+//  Created by Ronan O Ciosig on 5/6/21.
+//  Copyright © 2021 Sonomos.com. All rights reserved.
+//
+
+import Foundation
+
+public protocol DataProviding {
+    func catchPokemon()
+    func newSpecies() -> Bool
+    func pokemon(at index: Int) -> LocalPokemon
+}
